Add files via upload
This is a Math game that a dear colleague of mine worked on with me. The aim was to help people at any level to improve their quantitative prowess overall which would provide benefits across all areas of life. The core target market were children, however, people in general would find excellent use of such a program as Mathematics is omnipotent. Please view the documents for more detailed information.